requirements.txt
Django>=3.2,<4.0
gunicorn
dj-database-url
python-decouple
whitenoise
psycopg2-binary
djangorestframework